增加父关联查询逻辑
